+++ RELEASE_NOTES.md
@@ -6,7 +6,7 @@
 
 ## Upgrading
 
-<!-- Here goes notes on how to upgrade from previous versions, including deprecations and what they should be replaced with -->
+- The reporting API client version is updated to `>= 0.18.0, < 0.19.0`.
 
 ## New Features